Added JWT authentication and expense tracking

+++ src/main/java/com/example/splitwise/DTO/AddExpenseRequest.java
@@ -11,6 +11,6 @@ public class AddExpenseRequest {
     private Integer groupId;
     private Double amount;
     private Integer paidBy;
-    private List<Integer> members;
+    private List<Long> members;
     private String description;
 }
